moved qa.ts to lib folder
