+++ src/lib/qa.ts
